requirements.txt
assemblyai
numpy
torch
torchaudio
scikit-learn
scipy
pyannote.audio
pyaudio
pydub
langdetect